feat: 매칭 등록 후 매칭한 유저들끼리 확인 api 연결

+++ android/Signal/app/src/main/java/com/ongo/signal/data/model/match/MatchRegistrationResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class MatchRegistrationResponse(
     val latitude: Double,
-    val location_id: Int,
+    val location_id: Long,
     val longitude: Double,
-    @SerializedName("user_id") val userId: Int
+    @SerializedName("user_id") val userId: Long
 )
